Unify the canvases by moving groups around rather than using set_scrolling_region () (eliminates the double-draw during zoom on linux), fix verbose_time_cursor clamping, make fonts smaller on location dialog buttons, hopefully improve performance when dragging multiple regions on the same track, reduce some redrawing of clocks (still not done for samples clock), allow NLS=1 on OSX, other canvas performance improvements on OSX, build SAE-us.bindings (even though they are incomplete for now).
git-svn-id: svn://localhost/ardour2/branches/2.0-ongoing@3707 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/time_axis_view_item.cc
@@ -149,7 +149,7 @@ TimeAxisViewItem::init (const string& it_name, double spu, Gdk::Color& base_colo
 	vestigial_frame = new ArdourCanvas::SimpleRect (*group);
 	vestigial_frame->property_x1() = (double) 0.0;
 	vestigial_frame->property_y1() = (double) 1.0;
-	vestigial_frame->property_x2() = 2.0;
+	vestigial_frame->property_x2() = (double) 2.0;
 	vestigial_frame->property_y2() = (double) trackview.current_height();
 	vestigial_frame->property_outline_color_rgba() = ARDOUR_UI::config()->canvasvar_VestigialFrame.get();
 	vestigial_frame->property_fill_color_rgba() = ARDOUR_UI::config()->canvasvar_VestigialFrame.get();
